requirements.txt
django==4.1
python-dotenv==0.20.0
psycopg2-binary==2.9.3
djangorestframework==3.13.1
django-cors-headers==3.13.0
Pillow==9.2.0